chore: unnest websocket interceptor directory structure
