+++ src/interceptors/WebSocket/index.ts
@@ -1,32 +1,81 @@
-import { BatchInterceptor } from '../..'
-import type { WebSocketClientConnection } from './WebSocketClientConnection'
-import type { WebSocketServerConnection } from './WebSocketServerConnection'
-import { WebSocketClassInterceptor } from './implementations/WebSocketClass/WebSocketClassInterceptor'
+import { Interceptor } from '../../Interceptor'
+import { WebSocketClientConnection } from './WebSocketClientConnection'
+import { WebSocketServerConnection } from './WebSocketServerConnection'
+import { WebSocketClassTransport } from './WebSocketClassTransport'
+import { WebSocketOverride } from './WebSocketOverride'
 
-export type WebSocketEventsMap = {
+export type WebSocketEventMap = {
   connection: [
     args: {
       /**
-       * The connected WebSocket client.
+       * The incoming WebSocket client connection.
        */
       client: WebSocketClientConnection
 
       /**
-       * The original WebSocket server.
+       * The original WebSocket server connection.
        */
       server: WebSocketServerConnection
     }
   ]
 }
 
-export class WebSocketInterceptor extends BatchInterceptor<
-  [WebSocketClassInterceptor],
-  WebSocketEventsMap
-> {
+/**
+ * Intercept the outgoing WebSocket connections created using
+ * the global `WebSocket` class.
+ */
+export class WebSocketInterceptor extends Interceptor<WebSocketEventMap> {
+  static symbol = Symbol('websocket-class')
+
   constructor() {
-    super({
-      name: 'websocket',
-      interceptors: [new WebSocketClassInterceptor()],
+    super(WebSocketInterceptor.symbol)
+  }
+
+  protected checkEnvironment(): boolean {
+    // Enable this interceptor in any environment
+    // that has a global WebSocket API.
+    return typeof globalThis.WebSocket !== 'undefined'
+  }
+
+  protected setup(): void {
+    const webSocketProxy = Proxy.revocable(globalThis.WebSocket, {
+      construct: (
+        target,
+        args: ConstructorParameters<typeof globalThis.WebSocket>,
+        newTarget
+      ) => {
+        const [url, protocols] = args
+
+        const createConnection = (): WebSocket => {
+          return Reflect.construct(target, args, newTarget)
+        }
+
+        // All WebSocket instances are mocked and don't forward
+        // any events to the original server (no connection established).
+        // To forward the events, the user must use the "server.send()" API.
+        const mockWs = new WebSocketOverride(url, protocols)
+        const transport = new WebSocketClassTransport(mockWs)
+
+        // The "globalThis.WebSocket" class stands for
+        // the client-side connection. Assume it's established
+        // as soon as the WebSocket instance is constructed.
+        this.emitter.emit('connection', {
+          client: new WebSocketClientConnection(mockWs, transport),
+          server: new WebSocketServerConnection(
+            mockWs,
+            createConnection,
+            transport
+          ),
+        })
+
+        return mockWs
+      },
+    })
+
+    globalThis.WebSocket = webSocketProxy.proxy
+
+    this.subscriptions.push(() => {
+      webSocketProxy.revoke()
     })
   }
 }
